Remove environment variable based root directory detection
Having the project root directory be dependent on the NODE_ENV environment variable works fine when testing hof-bootstrap itself, but causes unexpected behaviour when testing apps which are dependent on hof-bootstrap.

Instead make the project root explicitly configurable, with a default of the current working directory.

Add a convenience method to the exported bootstrap module to allow for modifying global configuration. This allows for the default project root to be set in the integration tests without needing to be set in the constructor in each instantiation.