Fix translations
The previous implementation of translations didn't have the ability to support multiple languages because it did not specify the langauge anywhere in the translation method calls.

By moving the translations to middleware we can make use of built-in `Accept` header parsing to determine the user's desired language and translate accordingly (although we don't currently have content for any langauge other than `en`).

Additionally by automatically mounting translations from `hof-template-partials` we can simplify translation compilation in implementing projects as we remove any need to compile that content into translation files.

In particular this removes any need to define common translations in order to be able to load content for error pages and cookie/terms pages.