Add theme loader to bootstrap (#135)
* Add theme loader to bootstrap

Allows theme to either be passed as a module or as a string suffix.

* Allow (almost) any theme version

This will allow bootstrap to always inherit the version specified by an implementing project if one is specified. e.g. if a project wanted v3 of the theme to be used then it would not be overridden by this version.