Remove non-standard log levels
I'm not aware of any need for the custom levels, but they make trying to do logging in a predictable way difficult because the expected log levels don't exist and so logging `debug` or `verbose` level messages fails silently.

Remove the custom log levels and use the default levels instead.