Decouple custom csp directives and disabling csp
If an app needs to set custom directives then it cannot currently also leave the disabling of csp by setting `DISABLE_CSP` env var intact, so implementations need to implement this as well as it being implemented in `hof`, which is not ideal.

By splitting the disabling of csp from a top level `csp: false` property, to a sub-property of `disabled` then custom directives can be set, but will also be globally disabled independently, and the `DISABLE_CSP` environment variable is still respected.

Leaves the option to disable by setting `csp: false` for backwards compatability.