Bump hof-form-wizard major version
Depend on the latest major version of hof-form-wizard so that
fork evaluations now take into account inputs from previous
forms as well as the current one.

This change results in a new major version of hof since it
*could* cause problems if people have inputs of the same name
across multiple forms and are depending on those fields for
forking behaviour.