SD-50: Update winston logger
Currently applications using HOF customise their own loggers and ignore
the inbuilt functionality for logger that is in HOF. Currently the HOF
logger is using an out of date version of winston 2.x.x and is reliant
on Churchill for setting it up. This commit is to update winston to
3.x.x and setup the 7 logging levels with a log() function that used
to be in version 2.x.x before it became deprecated.

By implementing this, we should encourage HOF services to use the
inbuilt logger so we have consistency. This was initially setup and
tested on RIL (Refugee Integration Loan) so has been imported into HOF
so it can be more widely available.