Extend SASS config options in hof-build
The current implementation of hof-build doesn't allow for CSS output to be compressed or warnings to be suppressed (they cause a lot of noise in the console output).
This PR adds optional configuration to apply these 2 options via hof.settings, anyone not choosing to configure either of them will see the default behaviour ie. not compressed CSS output and deprecation warnings showing in the console output when compiling SASS.

Bump minor version

Update read me